field

+++ MAIN/models.py
@@ -21,7 +21,7 @@ logger = logging.getLogger(__name__)
 # Create your models here.
 class FileModel(models.Model):
     filename = models.CharField(max_length=300)
-    filesize = models.IntegerField(default=0)
+    filesize = models.BigIntegerField(default=0)
     key = models.CharField(max_length=300)
     uploaddate = models.DateTimeField(auto_now=True)
     browserable = models.BooleanField(default=True)        #设置文件是否对普通用户可浏览
